Refactor income tax calculation
Before this change, a giant global income tax rate was passed into the
simulation and that was applied to the cumulative annual income as a
simple multiplier.

This commit enhances that greatly:

* No more "global tax rate"
* Tax brackets introduced, based on 2024 IRS actuals
* Inflation simulation added to tax brackets
* Standard deduction introduced, also subject to inflation simulation

There are still a number of issues related to tax calculation and the
one drawback to this enhancement is giving a false sense of accuracy.
Next big tasks:

* Come up with a blanket estimate to simulate AGI
* Factor in capital gains