Ensure the returned plan has START of month values, not END
This is a subtle but important change. I didn't even really think about
it before, but after analyzing the output, it very much matters that
each row shows the start of the month. If the end of month values are
important -- just look at the beginning of the following month!