Introduce new class for reporting
This is a very rough, initial take. I'm not sure how this will shape up
but for now, I needed to get CSV export support back to help with unit
testing. I lifted the old code from the previous `retreat` project and
updated for schema changes, etc. It's working pretty well. Still just
not sure where it's going to go in the end.