Introduce a pseudo-unit test for ut05
This a large and complex task, really more functional testing than unit
testing, so I haven't really settled in on an approach. For now, it
runs two tests: 1) repeat until fixed period reach, 2) repeat until
assets depleted.

The only tests are to count the number of rows returned in the plan. But
it also calls the new Reporting class to dump out CSV and also dumping
out logs. This will definitely change as it matures. Just a first step.