Change interpretation of "null" second period to end-of-time
Prior, if it was missing, I set it to 1970-01-01, but that didn't work
for me. The new interpretation is that comparePeriods($p1, null) is
that null is in the far future and so null effectly always returns
a value of -1, because $p1 will always be before the end of time.