Refactor Simulation controller [m]
The controller was way too heavy: contained too much business logic
(which went from a theoretical problem to an actual problem when I
went to add a Symfony command for running one of the endpoints.

A new Service/Engine/Simulator class added to centralize the logic,
most of which was pulled from the old controller.