Finish rename of "plan" to simulation and "planEntry" to step [m]
No functional change, but topping off a slow-burning change as I came
to realize what I was actually doing. It's not a plan at all. It is what
it is: a simulation. And that became clearer than ever when I began
putting the appsmith app together and all my scenarios had to be tied
together into . . . something. A simulation!