Implement dependency injection [m]
A bit of an overhaul here, but here's the backstory: before cutting over
to Symfony, I wasn't using any framework: just POPO. And to get around
some limitations, LogFactory was introduced and I had to instantiate my
objects all over the place.

Well, now that I've gone over to the dark side...  ;)

This is now a Symfony app, so let's do it right. Here are the highlights:

* Got ride of LogFactory
* Added parameter bindings to services.yaml, for each env var in .env
* Massive pseudo "search and replace" for instances of "new Anything"
* Replaced with new constructors and DI everywhere

Note: this commit represents a very broken state. I'm in the middle of
a number of other large changes, but for that very reason, I wanted to
commit this subset of the overall current effort, just so it doesn't get
any worse than it already is.

The end result here is a much cleaner, more modern implementation. It
took some time, but I'll be better for it going forward.