Commit very large overhaul with respect to income
At first I wanted to break this down into discrete commits, as I usually
do, but this is so large and so sprawling, that there's really just no
point.

It began with the realization around a week ago that income tax, while
being calculated, was never being paid for. This began a week-long
journey of not only fixing that, but diving deeper into tax and setting
the stage for something more sophisticated. Discovered more issues here
and there and added new features there and here. Lastly, updated the
unit tests:

```
bash-5.1$ bin/phpunit
PHPUnit 9.6.16 by Sebastian Bergmann and contributors.

.......................................... 42 / 42 (100%)

Time: 00:04.121, Memory: 8.00 MB

OK (42 tests, 493 assertions)
```

And now I'm just going to commit this giant blob of updates. Bottom line
is: it's so much better. The Engine overhaul, the new Income and
IncomeCollection classes. Introduction of a SimulationParameters class.
More opportunities for DI and autowiring. I just feel I'm in a much
better spot now. It's really coming together and maturing.