Commit WIP for a real income tax engine
Note: this is broken and has a long way to go, but I wanted to commit
this first stab at it before moving forward, especially since I just
committed support for a tax engine indicator. Engine #1 is for a flat
effective tax rate to be used every year. Engine #2 (this commit) makes
an attempt to calculate a more realistic tax, based on income types.