Remove Money class and switch from floating point to integer math
Note: today is June 2, but this work was done February 22 and/or 23. I
had forgotten I'd worked on it and even more forgotten that I'd never
committed it. I ran a few checks and it seems to be okay but I can't say
that there aren't some issues lingering. This is, after all, an
extremely large change to the codebase.

Background: the floating point math (wrapped in the Money class) had
a large number of issues, especially with respect to passing objects
by reference. Meaning, what I thought were two different $money instances
were actually operating on the same data. Fixing this was going to take
far more work than was necessary, especially when fractions of dollars
are completely meaningless in this kind of simulation.

So all money is whole dollars from here on out. I have not yet looked at
the database impact. Everything there is still floats, so I'm hoping I
can get away with this float->int change in the code and leave the db
intact. I just don't have time right now to deal with that.