Add support for asset.ignore_until_year and .ignore_until_month
This ended up being a bit more complicated than I thought it would going
into it. But the concept here is the ability to introduce an asset at
a point in time after the simulation begins. Until that point in time
passes, it's as if it doesn't exist at all.

Before this change, the "begin_year" and "begin_month" only meant when
is it okay to start drawing from the asset. But the asset still existed
and (worse) was earning interest.

This allows a new asset to show up in the future. In the immediate case,
this is the sale of a house ten years down the road.

This commit is a WIP. Code is done but not yet tested.